test_requirements.txt
# Test-only dependencies are unpinned.
#
pip
coverage>=3.6
mock>=1.0.1
nose>=1.3.1
flake8
#
# Specify precise versions of runtime dependencies where possible.
PyYAML==3.10       # precise
